+++ CHANGES.md
@@ -1,5 +1,16 @@
 # @geoblocks/geoblocks changes
 
+## 0.2.3
+- Add utility functions.
+- In `BaseCustomizer`, the printExtent can be now set and get/set are dedicated methods.
+- `pdfA` (allow transparency) is now a spec.map optional param.
+- spec.attributes are now partial and `datasource` attribute is removed.
+- CreateSpecOptions accepts now every format.
+- Add a timeout and manage errors on the `getDownloadUrl` utils function.
+
+## v0.2.2
+- Add optional MVTEncoder
+
 ## v0.2.0
 - General refactor (rename classes / types)